Track allocation statistics in generated programs
To benchmark future changes, track the number of objects allocated
and the total amount dynamically allocated on the heap. This commit
introduces a glmalloc(size_t) function that tracks the running total
of memory allocated, and switches all other mallocs to use it.
alloc_obj() will now increment a tally for each object created.