Store only one copy of methods for builtins
The methods for each Float64, String, and Array object are the
same, so create and store only one copy of the data. The first
object created of each type creates the method structs and method
array, which is then stored in TYPE_Methods and reused for
subsequent objects.

Running against the compiler itself this change makes no difference
to the number of objects allocated, but saves 97MiB of memory on a
32-bit system.

        Objects:   Memory:
Before: 495,420    206MiB
After:  495,420    109MiB
--------------------------
              0     97MiB