Add infrastructure to build known-good revisions
A new directory, known-good, contains a script and makefile for building
the architecture-dependent known-good binaries for the required
revisions, when provided with an initial build from Parrot (or a script
invoking the compiler on Parrot). The main makefile will depend on these
revisions explicitly and cause them to be built when the requirement
changes.