Allow building from scratch without Parrot
Now ship the bitcode for minigrace-initial so only clang and LLVM are
required to start, rather than a copy of the Parrot version. gracelib
must be compiled for each architecture, but the bitcode of the compiler
itself is architecture-independent.

Provided clang and LLVM are installed, a fresh checkout needs only
"make" to build a functioning HEAD compiler. The intermediate versions
will be constructed along the way as required.