Generate exactly the needed number of params_x
LLVM code generation previously generated params_0..params_10 in every
method and the module initialisers, regardless of how many arguments
were used in the body. The generator will now generate only the
necessary number, including more than 11 if required. The array will be
stack-allocated as well, as it is not shared now.