Update method dispatch for delegative inheritance
User objects have a ->super pointer, which points at a
fully-instantiated object. When a method is not found on the current
object, dispatch is retried on that object. There are three cases for
the method so found:
1) A user method is invoked with self bound as the object the original
lookup was on.
2) A built-in method or method handling a var or def has
MFLAG_REALSELFONLY set in ->flags and is passed the physical object the
method was found on as self.
3) A built-in method which must virtually invoke other methods on self
has MFLAG_REALSELFALSO set in ->flags and is passed both the receiver
and the physical object the method was found on.