Remove (1) from multi-part names internally
Multi-part method names have had (1) (or (2), etc) embedded in their
names internally to allow arity checking. As multiple variadic arguments
will be required these numbers can no longer be determined at the call
site, and so must be removed. The typechecker will need to learn to deal
with arities after the parser is updated, but for the moment the test
case concerning this behaviour is disabled.