Add very basic gradual checks
Method parameters with named static types defined at the top level, or
built-in, will have the arguments in those positions type-checked
dynamically using the same machinery as pattern-matching (and with the
same limitations). Other parameters are not checked at all. A run-time
type error will terminate the program with a backtrace and report the
expected type, position, and parameter name.