Move -lm -ldl, etc to end of gcc command line
Some systems, notably Ubuntu 12.04, pass --as-needed to the linker by
default. On these systems, libdl and libm were not linked into the
executable as they appeared at the start of the command line and were
discarded as unused. Moving this part of the command line to the end
should fix the problem.