Partial fix to the self problem exposed in test 114.
The problem is that use of super assumed that the object passed to
inherits was unique to the calling object, and set the super object's
`self` value as the caller. In test 114, objects inherit from booleans,
which change the self value of the true and false literals, which causes
them to show erratic behaviour (true now thinks it is a different
object, for instance).

In order to fix this, I've had to remove the `self` field from the Grace
Object class, and instead required that every method exposed to Grace
take an additional parameter at the start which will be the `self`
value. I've gone through and added this to every method in the Java
standard libraries, but I've yet to update the code generated by
genjava.grace so a build now will obviously fail.

I've also renamed all of the GraceVoid values to GraceNothing, in line
with changes to the spec. This is also not yet reflected in genjava.

The invoke method has also been tweaked. It will now not fail when it
finds a method with the right name but that is not a real Grace method,
if a real Grace method exists elsewhere on the object. It also allows
method return types to be a subclass of GraceObject instead of strictly
GraceObject. I've updated the return types of methods to reflect this as
well. I haven't done this for GraceObject though, because extending
GraceObject classes can choose not to return the more specifc types (ie
may return a number from asString).