Big changes to the backend.
I've moved the static functionality that previously resided in
GraceObject into a new class, GracePrelude. The grace.lang classes that
require functionality contained therein all use import static to
retrieve the exact functionality they use. Grace modules extend
GracePrelude directly to get all of these methods for free, as before.
The classes in grace.lib now do this.

I've also added an egal field to GraceObject, and require a value for it
to be passed into the constructor or default to pointer equality. I've
yet to actually implement the equality other than pointer, partly
because I need to clarify precisely how to test the other kinds.

I'm about to rename everything to remove the Grace prefix. That will
appear right after this commit.