Working towards having the tests passing again.
Most of the changes are bugfixes in genjava and some of the runtime
libraries, but it's worth noting that there's currently a little hack in
compileCall to enforce direct calling of the print function. This can be
removed once print officially gets moved into the prelude (or it is
confirmed that it won't be).