Java backend functional again, now has egal.
Only the pattern matching tests remain. Egal uses reflection and so is
almost certainly terribly underperformant, but it passes the tests. Built
in classes will check their own equality and then defer to the egal
implementation.

I've had to change the way that the $outer variable was passed around.
The name was misleading, as in the scope where it was available it
actually equated to self, so there is now a self variable that refers to
the real object and a $self field which refers to the local object,
which may or may not be the value in self. There is also a $closure
field which refers to the nearest enclosing Java object. Outside of a
method this will be the same as $self, but in a method closure it will
refer to the object representing that closure.