Store modules by path to share initialisation
Previously, `import "x" as y` saved the initialised module in
"module_y", so it would not be shared with other imports with different
names and would conflict with other imports with the same name. Instead
track the mapping of imported name to path and always store the module
object in module_<path>, so there are no linking conflicts.