Ensure import path matches compilation path
The new import system allows modules to be imported from other
directories. The path used to import them must uniquely identify the
modules; it is possible to compile them from elsewhere and give them the
wrong build information, which will cause linking to fail later. The
compiler will now check the claimed path in the .gct file itself and
give a meaningful error when it does not match the import path.