Allow loading samples in JS UI
A new dropdown and button, similar to the test case system, permits the
user to select one of the code samples (currently all dialects) and load
them up. When they load a module depending on another (like a dialect),
a precompiled version of that dependency is fetched and loaded if the
module has not already been compiled. GCT files are also fetched and
stored in the cache.