Changes to enable better exception reporting in the ECMAScript backend.
genjs.grace changed to output moduleName information as well as lineNumber
information.  Globals are now updated using procedure calls (setLineNumber,
setModuleName) rather than direct assignments, so that these procedures can
in future be used as hooks for debugging.  Fixed irregularities in indentation.

gracelib.js updated to provide library support for writing to stderr, and to
implement the setLineNumber and setModuleName procedures.  Also added missing
methods on ExceptionPacket and Exception.  Added printBacktrace method for
compatibility with C backend.

minigrace.in.js now delegates the printing of the backtrace of an otherwise-
unhandled exception to the exception object, rather than duplicating this
function