Add 'imports' module providing resource loading
imports.loadResource(path) will map the extension on path to a handler,
which will be used to interpret the path into an object. Handlers are
registered using imports.registerHandler(extension, handler) where
handler is an object containing a loadResource(path) method. By default
a "txt" handler is defined, returning a string.