Make tarball Makefile update buildinfo.grace
The new search paths are set in Makefile.conf during configure, but must
be transferred to buildinfo.grace to be used by the compiler. The
tarball's makefile did not make this transfer and so the `make install`
target would misbehave. Update the buildinfo.grace file each time the
configuration changes.