Add new known-good to default generations cache
The version number is based on the generation of the commit built, which
is calculated during make. The generations.dat file notes several
revisions to short-circuit this calculation, and some versions are
included in git-calculate-generation directly for the use of systems
(notably Mac OS X) with old versions of bash.

This commit adds the current known-good version, generation 1926, to
both files as the previous included version was nearly 300 generations
ago and approaching the area where problems start to occur.