Peg known good at previous commit, host personally
The current tarball distribution refers to a commit that is not in this
repository's past, and so the tarball bootstrap could not possibly work.
Now the latest commit is marked as known-good and the tarball is hosted
on tim instead of mwh.

Now that the build can run in a sensible fashion, a Travis configuration
file has been added that tells the CI how to build and test the project.