Changes to Make process for dialects.
Changed genc to not link transitive imports
of StandardPrelude, since these are already linked
into gracelib.o.

Changed the dialects Makefile to capture more (all?)
of the dependencies between the dialect files. This
allows parallel compilation (e.g, with make -j5).

Changed the fsm dialect example not to raise an
exception, so that it can be run during Make.

Modified the --target imports option in the compiler
so that it outputs the imported file paths (rather
than crashing).  However, this isn't presently used in
the Makefile.

minigrace.js wsa not being used at all; index.html
used minigrace.in.js.  Now it uses minigrace.js,
which is greatly simplified.