Included primitiveArray.sort in gracelib.
The method is present in both the C and JS versions of gracelib.
Also includes a test (t154) and some other changes so that
a sortblock that fails to answer a number raises a TypeError.
(The test of the error message differs between C and JS).
Commented out ignored target in Makefile to avoid warning message.