Removed “else” from match()case().
match()case() does not have and else part, according
to the spec.  This commit comments out the else part parsing
clause.  The sat should be updated to remove the else part,
and the code generators not to support it, if this change proves
innocuous.

dialect.grace and staticTypes.grace were changed to remove
uses of else.