Runs js code for tests at the command line.
Uses node to set up an environment in which tests of the generates
JavaScript code can be run at the command line.  As of this commit,
only 53/127 pass.

To run the tests, cd to .../minigrace/js/tests and execute
    ./harness ../../minigrace . ""
This will compile the tests to JavaScript using the native compiler,
and then run the JavaScript using node and the usual test harness.

Most of the changes in this commit amount to cleaning up the JavaScript
so that it will run in strict mode.  While not necessary for node,
this makes finding JavaScript errors much easier.  The js code generator
also emits a postlude that makes the imported module function global
in a way that depends on the environment --- node or browser.