Makes all JS tests run under node.
Adds a command-line runner for grace code: js/tests/grace-node.

This involves quite extensive changes to the way that global
variables and global constants are declared, with the declarations
of the variables moving from gracelib.js to grace-node (for the
command line) and minigrace.in.js (for the browser).  Also adds
a makefile target (test.js) for running the tests at the command line.

Adds some new tests, and moves two known-failing tests to a
subdirectory.