Updates unicode tables to work with node.
unicode needs to be a global object; under node
this must be declared explicitly.   The program that
creates the table now add the declaration.   Note that
many of the changes in this file are due to the list of
categories being in hash-table order.   Others are due to new
unicode characters.