Adds “factory method” syntax.
Adds factory as a keyword, and makes factory method
equivalent to method returning an object.
Adds t163 for this syntax, and changes t030 to avoid use
of "factory" as an identifier.

The implementation of factory methods is entirely in
the parser,a nd is based on mwh's implementation of "constructor"s.