Parses and generates code for type expressions.
Adds new AST nodes typedef and typeliteral.  Type expressions
are represented as expression trees, which are evaluated at runtime
using the usual mechanism for expressions — method request.

Type operators &, | - and + are now defined in the StandardPrelude.
In the JS code, this necessitates that the methods in the
StandardPrelude are _added_ to Grace_prelude incrementally,
rather than when the import is completed.   Thus, StandardPrelude
is now imported specially in JS. _prelude is no longer inherited.
This simplifies the setup in the web browser and in the node.js test
harness. Adds new methods to gracelib.js to support types, and
removes dead code.

These changes have been incorporated into the C backend only as
far as necessary to support the new parse tree structure.  Features
not used by the compiler are not included in the C code generator,
nor in gracelib.c