Improves error messages for dialects.
Change exception control flow so that when a dialect
fails to load or crashes we get appropriate error messages,
distinct from when the dialect’s checker finds a problem in
the code.

Also adds a test (t161) for a dialect.
Puts staticTypes and requireTypes in the list of modules
loaded into the web browser on startup.