Removes Exception.raise from TypeVariant.
I don't understand why raising an Excpetion from a method
in the StandardPrelude should cause the C code generator
to generate bad code, but it does.  (The collectionsPrelude
raises many excpetions!) So, pro-tem, I indirectly raise
a NoSuchMethod exception instead.

The right solution for this is to re-define types so that the
method can actually be written.