Reverts compilation of import accessors.
Removes the generation of accessor methods for
named defined in import statements.  I had attempted to
treat them just like defs, but had evidently got this wrong,
so that the wrong object was being obtained from the closure
in some cases.  This should go back in when I can figure
out how to do it correctly.