Moves js/tests/t126 to knownFailing.
This test exercises the #ObjectInheritance option,
which allows inheritance from statle (non-fresh) objects.
ObjectInheritance works with the C code generator, but not with
the JS generator.

It would be nice to fix this, but it does not seem like a priority until
the language spec is changed to allow inheritance from stale objects.