Reorganizes sample modules.
Changes makefiles and positions of the samples in the file hierarchy.

The js Examples can't be run from the command line
until we have filesystem code for JS.  This commit removes them
from the default make target in sample/dialects

Also removes the js examples that require the dom module from the C
hierarchy.

Fixes the dynamic load method (in C) to correctly search
GRACE_MODULE_PATH.

Fixes a bug in which the built-in Block type was
not exported from gracelib.js.

Adds the filepath method to FileStreams in the io module.