Concatenate relevant JS files to form minigrace.js
The current setup for including the JavaScript output requires adding
each script to the web page separately. This change forms minigrace.js
by concatenating the files that would have otherwise been included
separately, so that only minigrace.js needs to be important. Irrelevant
scripts such as the node-server are excluded from this process, and the
tabs script is included separately as it is only relevant in the context
of the built-in editor. Including other scripts such as objectdraw in
the js directory will have it included in the concatenation.