More Makefile cleanup.
Touches l?/exists files only if they don’t exist.

Removes -XnoStrict (strict JavaScript is good!).
Perhaps Tim put these in to avoid having “use strict” in the
middle of a big concatenated file?

In the rule for MGSOURCEFILES:%.grace=%.gso, reinstates
old .gct file after making needed .gso files; it would be better
if the compiler did this itself, after checking that the old and new
files are equivalent, but this is OK as a temporary hack; it avoids
forcing the remake of the whole compiler because there is one
“new” .gct.