Fixes lexer when input contains \r
This commit also  eliminates the third copy of source (a very long
string).  Instead, lexinput takes a list of lines as its input,
from which all \r and \l characters have been removed.

It removes some unnecessary nesting and the dreadful
pragma that changes the visibility defaults, and adds
the necessary “is public” annotations.