Reverts attempt to compile outer directly.
Removes the reference to enclosing object, and reverts to
the use of the outer method (which fails in the case of inheritance).
TODO: I believe that this can be made to work by allocating the
register strings when they are first requested, not when the node
in question is compiled.