Moves file path utilities into separate module.
This is a small step towards making the modules cohesive.
But it’s also necessary because util.grace is not loaded into
the web IDE, so nothing defined there is available in that IDE.
The path in js/tests/grace-node ensures the the tests look
for dialects in both samples/dialects directories.