Include comment nodes in AST
With these changes, the parser no longer ignores comments. It
creates commentNodes out of them and then, if the comments are
surrounding a vardec, defdec, methoddec, typedec, class definition,
or methodtype, the comments are attached as children of that
corresponding node. If they aren't surrounding one of these nodes,
they are discarded and not attached anywhere in the AST. These
changes are necessary for GraceDoc, the Grace documentation
generator.