Cleans-up tests and adds new tests.
t179 was writing and re-reading a GCT file, but it was named
(without string representation) of a bug in the js file open
method.  Moreover, the output directory name
was not being set.  This commit exposes a method to set that
directory, and uses it in the test.

Adds t160, a second (failing) test for require_types. Renumbers
string_test so that the two requireTypes tests can be adjacent.
Tweaks js/tabs.js to recognize the checker failure as a compilation
failure.