Harness_js takes pathnames relative to invoker.
Previously, all of the harnesses required their arguments to be
relative to the execution directory.  Also make execution directory
optional (defaulting to the directory of the harness).

All the js harnesses now use js/grace as the executor, by providing it
with a suitable GRACE_MODULE_PATH.  js compilation output goes in
the js directory, so as not to overwrite .gct files in the modules
directory.

TODO: the other harnesses should also accept paths relative to the
invoker.