Keeps value of try-catch expressions.
Try-catch _expressions_ now work in the JS backend.
They also work in the C backend, modulo one issue:
if the finally block *returns*, the program crashes with error 11.
This is due to the block-return logic (which used longjump) clashing with
the exception-handling logic (which also uses long jump).
I think that the try-finally never gets to clean up its state because
the block return whisks control-flow away from it.

Adds test t185 for the various cases.