Replaces `this` by `superDepth` in JS lists.
The native code inserts for list objects were manipulating `this`.
In the presence of inheritance, `this` does not have the data
attributes that were assumed.  However, the variable `superDepth`
holds the super object on which the attributes do exist, so replacing
`this` by `superDepth` seems to make the native code work when
inherited.