Implements type-checks on block arguments.
If there is a single argument, then the block is treated as a pattern
and matched to the argument.  If there are multiple arguments, the
types are saved in a property of the block, and the block.apply method
checks them.  Because match uses applyIndirectly, this check is bypassed
when the block is applied as a side-effect of a request on match.

This comit primarily addresses the JS backend.  Changes to the C
library are confined to generating the correct exception, since the
checks were already implemented.  However, with the C backend, they
seem to work inconsistenly.