Corrects test for column offset of `class`.
The test for ast line numbers introduced in 433d6fdc was actually
wrong, in that the column for the class keyword should have been 1,
not 2. Changes to the lexer fixed this; this commit fixes the test.
It also removes the test from those from the C backend, because
the C backend crashes on it and I don't have the time to maintain it.