Updates build of tarball.
Files math.gct and math.gcn were listed explicitly in the recipe for
target `c`.  This is used in make tarball. Now references modules/math.
Adds modules directory to tarball recipe.  Populates it
with unixFilePath, since this module is  used in the compiler.
Changes configure to include unicodeFilePath in STATIC_MODULES.