Updates Makefile for new KG compiler.
Now that the $(KG) compiler understands the dir argument,
we can get rid of a lot of the file copying.

There seems to be a bug in make (or a limitation in my understanding)
through which pattern rules like
    collectionsPrelude%gct collectionsPrelude%gcn:
                    collectionsPrelude.grace l1/minigrace
induces a dependency of l1/collectionPrelude.{gct,gcn} on l1/minigrace.
Since l1/minigrace depends on l1/collectionPrelude.{gct,gcn}, make then
complains about a circular dependency.

    make: Circular l1/collectionsPrelude.gct <- l1/minigrace dependency.
    make: Circular l1/collectionsPrelude.gcn <- l1/minigrace dependency.

So rules like the above are commented-out in the Makefile, and replaced
with a rule for only the .gct file.