Limits number of errors to rerun (default 10).
gUnit now has a public variable, numberOfErrorsToRerun, that
allows the programmer to limit the number of erring tests to rerun.
The minitest dialect allows this variable to be modified using a pair
of eponymous methods.

For minitest, the stack backtrace is limited in extent using the name
of the approriate minitest method, since th etest name does not appear
on the callstack.