Optimizes emptySequence; removeAll tries harder.
Removes unnecessary type-checks, which cause
infinite recursion.  This is because SuccessfulMatch contained
a list, and creating the list did a type-check.
Avoids native-code lists when list is inherited.

If one of the arguments to removeAll()ifAbsent is not present, the
others are removed, unless the absent action terminates the method.
Updates tests for remove()ifAbsent to test the new semantics.

Fixes exception messages on bounds-errors in ranges.