Add makefile targets to check js code.
It’s easy to create bad JavaScript.  Running jsl (JS lint)
on it is a good idea.  These makefile targets make it a tiny bit
easier. jsl must already be installed.
checkjs: checks gracelib.js
checkgenjs: creates ast.js if necessary, and checks it as a sample
	    generated file.
The config file for genjs is actually specific to ast.grace, because it
+defines the two modules that ast imports.