Removes debugging printouts from t001.
I think that I found the reason that this test was crashing —
missing gc guards in gracelib.c that showed up when running
in a try-catch.  At least, the problem has not manifested itself
for a while, so it’s time to take out the debugging printouts.  It
seems that the tests are not the culprit.