Adds methods to C-library primitive lists.
Fairly easy to do, even if it is of questionable value.
I was tired of writing correct Grace code in the compiler,
only to have it fail because of a method missing on
primitive lists.
Also replaces the few requests to reduce by fold()startingWith.
Creates test suite for C lists and updates suite for JS lists.