Detects class declarations in illegal places.
The paerser accepts (undotted) class declarations inside methods,
because dotted class declarations are legal there, and the same
code path is used for both.  But undotted class declarations are
method declarations, and are legal only inside an object.  This commit
checks that all methods are inside an object, but the only ones that
might fail the check are methdods created by class declarations—
other methods would have been rejected by the parser.