Reverts varargs to primitiveLists.
This reverts those parts of commit 9fd01b1a that made
variable-arity methods create sequences out of their arguments.
They now create primitiveLists instead.   This was necessary
because the sequence method itself (along with list and dictionary)
is (temporarily) a variable-arity method, and must be defined
before sequence becomes available in the standard prelude for
the generated code to access.

Why is sequence temporarily a variable-arity method?  To allow
sequence [1, 2, 3] and sequence.with(1, 2, 3) to coexist during
the migration.