Refactors parsing of object constructors.
Classes, factory methods and objects all have their separate
parsing routines, in spite of the fact that they are the same
syntactic construct.  This introduces a new parsing routine for
objects, and refactors the other three to use it.
It also modifies the AST so that inheritsNodes are stored
_only_ in the superclass and usedTraits fields of the object,
and not also redundantly stored in the body.