Re-initializes util.lines for each new compile.
`lines` should start as an empty list; the lexer appends lines to it.
Although the real util module makes `lines` a constant list,
and initializes it each time the compiler is loaded, the fake
util module in gracelib is persistent from one compile to
another.  So, we must manually re-initialize the lines constant.