Preseves modification times on install.
Install was making the .gct and .js appear older than the .grace.
Moves warning about paths not containing the install location
into a script, because some platforms seem to use /bin/sh to
run make commands, and [[ ... ]]] tests are not understood.
Adds warning for PATH as well as GRACE_MODULE_PATH.