Changes directory before running the l1 compiler.
This, together with removing "./" from the search path,
seems to fix the problem of minigrace linking with the
wrong libraries.   unixFilePath was still being picked from the
current (l1) directory rather than from ..  Why only this file?
I was unable to find a command that both ran the l1 compiler
so that it found the correct unicode.gso, and so that it linked
the generated feels with the correct unixFilePath.gcn
Hence, this version of the makefile separates compilation and
linking.