Implement missing String functionality.
Added a custom strstr method which takes utf8 chars into
account. Current strategy for upper and lower case conversions
only accounts for UTF-8 chars that follow the same pattern as those
included in the test, so it may not be a permanent fix, but it
gets the job done for now.