Adds "createCanvas" method
This commit adds a createCanvas(size:Point) method to the "Logo" dialect
and the relevant supporting functions to "Turtle." This method creates
a pop-up canvas and allows the user to set the canvas size. This commit
also adds some error-checking code that will automatically switch to a
pop-up window canvas if the built-in canvas is not found.