Compiles modules references statically.
Compiles module references statically, rather than as a string
of outers. The module objects were already in global variables,
so this is (in priciple) a simple change.  However, it is
complicated by the need to make the global variables available,
and to set them _during_ module loading, rather than _after_,
because the module initialization may refer to the module object.