Streamlines callmethod.
With the removal of superobjects, we can simplify callmethod.
This improves call times significantly.  Avoiding metaprogarmming
is a big win; this version consequently limits callmethod to 10
arguments.  It also avoids checks on the normal path, instead
using exception handling to catch the errors and figure out what
went wrong afterwards.