Omits accessors for top-level confidential fields.
Not clear why they were ever being generated!  We could
flag them as confidential, but since modules can't be inherited,
only local code can access them, and that code uses direct
references to the variables.  So an accessor can never be used.

This commit also removes code to implement the `parent` annotation,
since it was intertwined with the other code being altered.