Eliminates blankNodes.
It's unclear why they ever existed.  The location of each symbol
is known, so the location of blank lines can be inferred if needed.
The presence of blankNodes, for example, in a method body, disrupted
the chaining of return values, because blank nodes do not evaluate
to done, but to whatever value precedes the blank line.

Updates two tests; t180, to remove the blankNodes from the pretty-
printed ast, and t183, to change the way that blank lines are checked
for their influence on comment placement.