Recognizes module()object and prelude as outer.
The special case of identifier "module()object" is now classified
as an outer, as is prelude. Perhaps prelude and module should be cases
of outerNode. In the toGrace(_) methods on callNodes and
memberNodes, self and outer are dropped.  This is not always
right, but it's right more often than wrong.