Adds tests for lexer; fixes lexer.
Deletes a now-redundant failing-test of indentation change
of one. A string interpolation that spans lines is now clearly
prohibited by the lexer.  Formerly, the lexer allowed it, but it
would fail to parse, with a confusing error message relating
to the left-paren that the lexer had inserted for the interpolation.