Visits type parameters exactly once.
Type parameters to methods were duplicated: in the method node
and also in the first signature part node.  Now they are stored only
in the signature part node.  Visitors now visit the signature parts and
the type parameters.
Syntax errors in standardGrace and collectionsPrelude revealed by
this correction are repaired.