Adds type cache to &, | and + types.
Uses a js object property lookup rather than a
Grace fastDict or hashMap.  There are a coupel of reasons for this.
Because standardGrace is compiled in a special way, importing fastDict
doesn't work.  The error is: UninitializedVariable: attempt to read
uninitialised variable fd, where fd is the nickname.  Also, the classUid
is a JS string, and converting it to a Grace string is expensive, as is
computing its hash — this was a problem with hashMap too.

Using a js object property lookup avoids unnecessary Grace object creations,
as well as capitalizing on js arrays being hash tables with string keys.