Collection cleanup
Adds ≠ to Collection, but not hash.  Sequence includes
hash, but the mutable collections do not, to stop them from
inadvertently being used as keys in dictionaries.  A new type
Sequenceable is introduced as a supertype of Sequenceable collections.
The method asDictionary is removed, since the `as...` methods
introduce cross-dependencies.  The tests are changed to reflect the
above