Imports module with relative path; resolves #295
Submodules are imported with paths relative to
the importing module.  This fix allows the compiler
to find the compiled code when it is created in the
same folder as the source.  The change is to index the
gct cache by module name, rather than path name.
This means that it is impossible to import two modules
with the same name, but that restriction is also imposed
by the gracecode_‹modulename› naming convention for compiled
code (#298).