Eliminates SHA computation when in browser.
The problem of using the wrong JS file that the SHA is intended
to solve does not exist in the browser.  Since the JS library that
we used to compute the SHA runs on a file, and thus won't work
in the browser, the simple solution is to eliminate the SHA computation.

We also eliminate the inclusion of source in the compiled code in the
broswer; it's not necessary, since the source is visible in the editor.