Fixes bugs in error reporting for case.
When multiple cases match, we must report an error.
When the error occurs in module initialization,  there
is no method on the top of the request stack;  attempting
to print the undefined method's name caused a crash.
Cases in a match(_)case(_) are now referred to as case 1,
case 2, etc, not case 0, case 1...