Removes -j flag from Makefile
There is a window between the compiler creating its
output file and writing to it that can cause parallel
compilations to fail, if another compilation seeking the gct
looks at the output file at the wrong time.  Output files
should be created atomically by writing to a temp file and
then renaming it.