Don't compile certain methods as accessors
If it's a once method, it should not be compiled as
an accessor.  If the method has a declared return type,
we need the type check, so we don't compile it as an accessor.
if the method has a parameter (issue #286), it's not an accessor.
TODO: check to see if the return type conforms to the type of
the variable, in which case the type check is unnecessary, and an
accessor with no check would be OK.

Adds tests to t020