Distinguishes dialects from the prelude
This is a transitional version of the compiler that is needed
to compile a version of minigrace that defines intrinsics and
the standard dialect cleanly.  In the prior version of minigrace,
standardGrace was treated specially, and contained both built-in
definitions from gracelib, and those actually present in the
standardGrace.grace module.