Uses sha256 hashes to find compiled code
This change should avoid repeated recompilation.  Caches
the location of code modules to avoid repeated searches;
this cache should probably be moved to the module AstNode.
In the browser, simply looks for the corresponding compiled
module function, rather than pretending that there is an object
file. Removes check for transitive imports, since the compiler
doesn't need them.  The executable will fail to load them at
runtime if they can't be found at that time — but this can
happen anyway.