Eliminates treatment of standardGrace as prelude
Reomves the confusion between "prelude" and "standardGrace".
Annotations must be declared, and compilation of annotation
defs is fixed.  The suggestion mechanism is no longer used to
suggest alternatives to an undeclared variable; instead, the
alternatives are part of the error message.  This makes the
output more compact, and visible in the IDE, where the suggestions
are not.  Dependencies are updated to reflect the splitting of
the standard prelude into multiple modules.