Refactors gct writing into identifierresolution
The "secret" of the gct format was formerly split between
xmodule, which wrote it, and identifierresolution, which read it.
Such a secret should be localized in a sigle module; xmodule might
seem to be the obvious choice, but reading the gct requires
creating symbol tables, which is the perrogative of identifier
resolution.

Eventually, the gct code ought to be subsumed into a scope module,
since the purpose of the gct is to save and restore scope information
across separate compilations.