Removes old compiled files before running tests
The scenario that makes this necessary is as follows.  Collections has
been modified, but sequenceTest has not, so the old compiled file
js/tests/sequenceTest.js is still valid.  However, the old compiled
file requires, and forces the loading of, an old compiled version of
collections.js.  When a test that imports sequenceTest is run, it
requires standard, which requires the current version of collections.
Now sequenceTest and standard require incompatible versions.