Adds artificial limit to gct generation.
Sometimes gct generation seems not to terminate.  Since this code is
going to be replaced anyway, I just added a bound on the number of
paths. I think that there is also a slowdown when sets get large,
and this may be the cause of the problem, or at least an exacerbation.
Sets should be reimplemented using maps.