Multiple fixes to gracelib
value_isMe, num_identityHash and string_identityHash
are functionally identical to == ad hash, but need to be
implemented in distinct functions so that they can be labelled
as confidential.  Some error-handling code was making (incorrect)
assumptions about the internal structure of objects, rather than
requesting Grace methods on those objects.